+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Upwell structures on grid to D-Scan result page
 - Deployables on grid to D-Scan result page
 - Pos / Pos Modules on grid to D-Scan result page
+- Destination system to Ansiblex Jump Gate in D-Scan
 
 ### Changed